Support reassembly from split metrics
Add a configuration key to define the input
format for the topic DustDevil consumes. If
the format is `split', then the split metrics
are reassembled into batches before posting
them.

This way it is possible for DustDevil to
integrate derived metrics calculated by Hurricane
into the metric batches prior to forwarding
them.